+++ package/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.2.1] - 2021-07-07
+- Fix ``OnWillEnablePatch``
+- Add Open Harmony Log button when log file exists
+- Remove ability to hide inactive patches
+- Use static constructor to initialize patch manager
+
 ## [1.2.0-exp.1] - 2021-04-22
 - Fix: ``hideInactive`` does also hide groups when all patches are inactive
 - Fix: 2019.4 compatibility ``hasFocus``